Validate non-iterating systems for family hook illegal interfaces. Merge family hook system tests to the main system test file.
