+++ src/commonMain/kotlin/com/github/quillraven/fleks/world.kt
@@ -187,6 +187,19 @@ class WorldConfiguration(@PublishedApi internal val world: World) {
      */
     private fun setUpAggregatedFamilyHooks() {
 
+        // Validate systems against illegal interfaces.
+        world.systems.forEach { system ->
+            // FamilyOnAdd and FamilyOnRemove interfaces are only meant to be used by IteratingSystem.
+            if (system !is IteratingSystem) {
+
+                if (system is IteratingSystem.FamilyOnAdd)
+                    throw FleksWrongSystemInterfaceException(system::class, IteratingSystem.FamilyOnAdd::class)
+
+                if (system is IteratingSystem.FamilyOnRemove)
+                    throw FleksWrongSystemInterfaceException(system::class, IteratingSystem.FamilyOnRemove::class)
+            }
+        }
+
         // Register family hooks for IteratingSystem.FamilyOnAdd containing systems.
         world.systems
             .mapNotNull { system -> if (system is IteratingSystem && system is IteratingSystem.FamilyOnAdd) system else null }
